Changed debugging, buffer changes, etc. Prints for debugs instead create signals Buffers have been changed significantly Sending no longer makes a thread due to improvements
Buffers had issues with binary data in streams and would produce garbage data when reading, breaking table deserialization. Writing to a stream, however, is perfectly fine.

Work still needs to be done to keep a socket open for reuse to prevent the level of delay it causes and occasional deadlocks.